Feat: 게시글 연관 속성 추가 #141

+++ src/main/java/com/nhnacademy/marketgg/server/service/impl/DefaultCategoryService.java
@@ -11,12 +11,13 @@ import com.nhnacademy.marketgg.server.exception.categorization.CategorizationNot
 import com.nhnacademy.marketgg.server.exception.category.CategoryNotFoundException;
 import com.nhnacademy.marketgg.server.repository.categorization.CategorizationRepository;
 import com.nhnacademy.marketgg.server.repository.category.CategoryRepository;
+import com.nhnacademy.marketgg.server.repository.customerservicepost.CustomerServicePostRepository;
+import com.nhnacademy.marketgg.server.repository.product.ProductRepository;
 import com.nhnacademy.marketgg.server.service.CategoryService;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -26,6 +27,8 @@ public class DefaultCategoryService implements CategoryService {
     private final CategorizationRepository categorizationRepository;
     private final ElasticProductRepository elasticProductRepository;
     private final ElasticBoardRepository elasticBoardRepository;
+    private final ProductRepository productRepository;
+    private final CustomerServicePostRepository customerServicePostRepository;
 
     @Transactional
     @Override
@@ -45,7 +48,8 @@ public class DefaultCategoryService implements CategoryService {
     }
 
     @Override
-    public List<CategoryRetrieveResponse> retrieveCategoriesByCategorization(String categorizationId) {
+    public List<CategoryRetrieveResponse> retrieveCategoriesByCategorization(
+            String categorizationId) {
         return categoryRepository.findByCategorizationCode(categorizationId);
     }
 
@@ -68,9 +72,12 @@ public class DefaultCategoryService implements CategoryService {
     @Transactional
     @Override
     public void deleteCategory(final String id) {
-        Category category = categoryRepository.findById(id)
-                                              .orElseThrow(CategoryNotFoundException::new);
+        Category category = categoryRepository.findById(id).orElseThrow(CategoryNotFoundException::new);
 
+        if (productRepository.existsByCategory(category) ||
+                customerServicePostRepository.existsByCategory(category)) {
+            return;
+        }
         categoryRepository.delete(category);
         elasticProductRepository.deleteAllByCategoryCode(id);
         elasticBoardRepository.deleteAllByCategoryCode(id);
